requirements.txt
Pillow==10.3.0
pillow_heif==0.16.0
streamlit==1.27.2
